server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.1
pymysql>=1.0.0
python-dotenv>=0.19.0
apscheduler>=3.8.0
python-telegram-bot>=13.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
requests>=2.26.0
cryptography>=3.4.8